Some refactoring in Permissions handlers
- Add default implementations of playerAddTransient() &
playerRemoveTransient() to the Permissions superclass.  "Pure"
superperms handlers (e.g. bPermissions, PermissionsBukkit) use those
methods, while handlers which use their own API can override them.

- Moved log and plugin fields to the Permissions superclass, since
they're present in all handlers.
